+++ VaxineApp/VaxineApp/ViewModels/AppShellViewModel.cs
@@ -4,8 +4,6 @@ using System.Collections.Generic;
 using System.Text;
 using System.Windows.Input;
 using VaxineApp.ViewModels.Base;
-using VaxineApp.Views.AboutUs;
-using VaxineApp.Views.Feedback;
 using VaxineApp.Views.Help;
 using VaxineApp.Views.Home;
 using VaxineApp.Views.Home.Profile;
@@ -13,7 +11,6 @@ using VaxineApp.Views.Login;
 using VaxineApp.Views.Settings;
 using Xamarin.Essentials;
 using Xamarin.Forms;
-using VaxineApp.Views.PrivacyPolicy;
 using DataAccessLib.Databases;
 using Newtonsoft.Json;
 using DataAccessLib.Models;
