+++ VaxineApp/VaxineApp/ViewModels/Settings/Main/SettingsViewModel.cs
@@ -3,13 +3,13 @@ using System.Collections.Generic;
 using System.Text;
 using System.Windows.Input;
 using VaxineApp.ViewModels.Base;
-using VaxineApp.Views.AboutUs;
-using VaxineApp.Views.Feedback;
-using VaxineApp.Views.PrivacyPolicy;
+using VaxineApp.Views.Settings.AboutUs;
 using VaxineApp.Views.Settings.AppUpdates;
+using VaxineApp.Views.Settings.Feedback;
+using VaxineApp.Views.Settings.PrivacyPolicy;
 using Xamarin.Forms;
 
-namespace VaxineApp.ViewModels.Settings
+namespace VaxineApp.ViewModels.Settings.Main
 {
     public class SettingsViewModel : BaseViewModel
     {
