+++ VaxineApp/VaxineApp/Views/Home/Area/Influencer/AddInfluencerPage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="VaxineApp.Views.Home.Area.Influencer.AditInfluencerPage"
+             x:Class="VaxineApp.Views.Home.Area.Influencer.AddInfluencerPage"
              xmlns:ViewModel="clr-namespace:VaxineApp.ViewModels.Home.Area.Influencer"
              xmlns:local="clr-namespace:VaxineApp.Renderers"
              Shell.TabBarIsVisible="False"
