+++ VaxineApp/VaxineApp/Views/Home/Area/School/AddSchoolPage.xaml
@@ -1,30 +1,20 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="VaxineApp.Views.Home.Area.Masjeed.AditMasjeedPage"
-             xmlns:ViewModel="clr-namespace:VaxineApp.ViewModels.Home.Area.Masjeed"
+             x:Class="VaxineApp.Views.Home.Area.School.AddSchoolPage"
+             xmlns:ViewModel="clr-namespace:VaxineApp.ViewModels.Home.Area.School"
              xmlns:local="clr-namespace:VaxineApp.Renderers"
              Shell.TabBarIsVisible="False"
-             Title="Add Masjeed">
+             Title="Add School">
   <ContentPage.BindingContext>
-    <ViewModel:AddMasjeedViewModel></ViewModel:AddMasjeedViewModel>
+    <ViewModel:AddSchoolViewModel></ViewModel:AddSchoolViewModel>
   </ContentPage.BindingContext>
   <ContentPage.Content>
     <StackLayout Padding="40" Spacing="10">
-      <Entry Placeholder="Masjeed Name"
-             Text="{Binding MasjeedName}"></Entry>
+      <Entry Placeholder="School Name"
+             Text="{Binding SchoolName}"></Entry>
       <Entry Placeholder="Key Influencer"
              Text="{Binding KeyInfluencer}"></Entry>
-
-      <StackLayout Orientation="Horizontal" Spacing="10" Padding="0, 10, 0, 0">
-        <CheckBox IsChecked="{Binding DoesImamSupportsVaccine}"></CheckBox>
-        <Label Text="Does Imam Supports Vaccine?"></Label>
-      </StackLayout>
-
-      <StackLayout Orientation="Horizontal" Spacing="10">
-        <CheckBox IsChecked="{Binding DoYouHavePermissionForAdsInMasjeed}"></CheckBox>
-        <Label Text="Do You Have Permission For Ads In Masjeed?"></Label>
-      </StackLayout>
       <StackLayout Orientation="Horizontal">
         <Entry Placeholder="Latitude"
                IsEnabled="False"
@@ -50,7 +40,7 @@
               WidthRequest="130"
               HeightRequest="50"
               HorizontalOptions="Center"
-              Command="{Binding SaveMasjeedCommand}"></Button>
+              Command="{Binding SaveSchoolCommand}"></Button>
     </StackLayout>
   </ContentPage.Content>
 </ContentPage>
