+++ VaxineApp/VaxineApp/Views/Settings/AboutUs/AboutUsPage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="VaxineApp.Views.AboutUs.AboutUsPage"
+             x:Class="VaxineApp.Views.Settings.AboutUs.AboutUsPage"
              Shell.TabBarIsVisible="False"
              Title="About us">
   <ContentPage.Content>
