+++ VaxineApp/VaxineApp/Views/Settings/Feedback/FeedbackPage.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="VaxineApp.Views.Feedback.FeedbackPage"
-             xmlns:ViewModel="clr-namespace:VaxineApp.ViewModels.Feedback"
+             x:Class="VaxineApp.Views.Settings.Feedback.FeedbackPage"
+             xmlns:ViewModel="clr-namespace:VaxineApp.ViewModels.Settings.Feedback"
              Shell.TabBarIsVisible="False"
              Title="Feedback">
   <ContentPage.BindingContext>
