+++ VaxineApp/VaxineApp/Views/Settings/PrivacyPolicy/PrivacyPolicyPage.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="VaxineApp.Views.PrivacyPolicy.PrivacyPolicyPage"
-             xmlns:ViewModel="clr-namespace:VaxineApp.ViewModels.PrivacyPolicy"
+             x:Class="VaxineApp.Views.Settings.PrivacyPolicy.PrivacyPolicyPage"
+             xmlns:ViewModel="clr-namespace:VaxineApp.ViewModels.Settings.PrivacyPolicy"
              Shell.TabBarIsVisible="False"
              Title="Privacy Policy">
   <ContentPage.BindingContext>
