+++ VaxineApp/VaxineApp/Views/Settings/PrivacyPolicy/PrivacyPolicyPage.xaml.cs
@@ -7,12 +7,12 @@ using System.Threading.Tasks;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
-namespace VaxineApp.Views.Home.Area.School
+namespace VaxineApp.Views.Settings.PrivacyPolicy
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
-    public partial class AditSchoolPage : ContentPage
+    public partial class PrivacyPolicyPage : ContentPage
     {
-        public AditSchoolPage()
+        public PrivacyPolicyPage()
         {
             InitializeComponent();
         }
